+++ Realm/Realm/Realm.cs
@@ -688,7 +688,7 @@ namespace Realms
         }
 
         /// <summary>
-        /// Factory for a write <see cref="Transaction"/>. Essential object to create scope for updates.
+        /// Begins a write transaction for this Realm.
         /// </summary>
         /// <example>
         /// <code>
@@ -708,6 +708,7 @@ namespace Realms
             ThrowIfDisposed();
             ThrowIfFrozen("Starting a write transaction on a frozen Realm is not allowed.");
 
+            SharedRealmHandle.BeginTransaction();
             return new Transaction(this);
         }
 
@@ -786,6 +787,47 @@ namespace Realms
         }
 
         /// <summary>
+        /// Asynchronously begins a write transaction for this Realm.
+        /// </summary>
+        /// <remarks>
+        /// This method asynchronously acquires the write lock and then dispatches the continuation on the original
+        /// thread the Realm was opened on. The transaction can then be committed either asynchronously or
+        /// synchronously.
+        /// <br/>
+        /// When invoked on a thread without SynchronizationContext (i.e. typically background threads), this method
+        /// calls <see cref="BeginWrite"/> and executes synchronously.
+        /// </remarks>
+        /// <example>
+        /// <code>
+        /// using (var trans = await realm.BeginWriteAsync())
+        /// {
+        ///     realm.Add(new Dog
+        ///     {
+        ///         Name = "Rex"
+        ///     });
+        ///     await trans.CommitAsync();
+        ///     // or just
+        ///     // trans.Commit();
+        /// }
+        /// </code>
+        /// </example>
+        /// <returns>An awaitable <see cref="Task"/> that returns a transaction in write mode.
+        /// A transaction is required for any creation, deletion or modification of objects persisted in a <see cref="Realm"/>.</returns>
+        public async Task<Transaction> BeginWriteAsync()
+        {
+            ThrowIfDisposed();
+            ThrowIfFrozen("Starting a write transaction on a frozen Realm is not allowed.");
+
+            if (!AsyncHelper.HasValidContext)
+            {
+                return BeginWrite();
+            }
+
+            await SharedRealmHandle.BeginTransactionAsync();
+            return new Transaction(this);
+        }
+
+        /// <summary>
         /// Execute an action inside a temporary <see cref="Transaction"/> on a worker thread, <b>if</b> called from UI thread. If no exception is thrown,
         /// the <see cref="Transaction"/> will be committed.
         /// </summary>
@@ -819,13 +861,48 @@ namespace Realms
         /// Action to execute inside a <see cref="Transaction"/>, creating, updating, or removing objects.
         /// </param>
         /// <returns>An awaitable <see cref="Task"/>.</returns>
+        [Obsolete("Use Realm.WriteAsync(Action action) instead.")]
         public Task WriteAsync(Action<Realm> action)
         {
+            ThrowIfDisposed();
+
             Argument.NotNull(action, nameof(action));
 
             return WriteAsync(tempRealm =>
             {
                 action(tempRealm);
+                return true;
+            });
+        }
+
+        /// <summary>
+        /// Execute an action inside a temporary <see cref="Transaction"/>. If no exception is thrown, the <see cref="Transaction"/> will be committed.
+        /// <b>If</b> the method is not called from a thread with a <see cref="SynchronizationContext"/> (like the UI thread), it behaves synchronously.
+        /// </summary>
+        /// <example>
+        /// <code>
+        /// await realm.WriteAsync(() =&gt;
+        /// {
+        ///     realm.Add(new Dog
+        ///     {
+        ///         Breed = "Dalmatian",
+        ///     });
+        /// });
+        /// </code>
+        /// </example>
+        /// <param name="action">
+        /// Action to execute inside a <see cref="Transaction"/>, creating, updating, or removing objects.
+        /// </param>
+        /// <returns>An awaitable <see cref="Task"/>.</returns>
+        public Task WriteAsync(Action action)
+        {
+            ThrowIfDisposed();
+
+            Argument.NotNull(action, nameof(action));
+
+            return WriteAsync(() =>
+            {
+                action();
                 return true;
             });
         }
@@ -858,6 +935,7 @@ namespace Realms
         /// </param>
         /// <typeparam name="T">The type returned by the input delegate.</typeparam>
         /// <returns>An awaitable <see cref="Task"/> with return type <typeparamref name="T"/>.</returns>
+        [Obsolete("Use Realm.WriteAsync(Func<T> function) instead.")]
         public async Task<T> WriteAsync<T>(Func<Realm, T> function)
         {
             ThrowIfDisposed();
@@ -891,6 +969,43 @@ namespace Realms
             }
 
             return (T)result;
+        }
+
+        /// <summary>
+        /// Execute a delegate inside a temporary <see cref="Transaction"/>. If no exception is thrown, the <see cref="Transaction"/> will be committed.
+        /// <b>If</b> the method is not called from a thread with a <see cref="SynchronizationContext"/> (like the UI thread), it behaves synchronously.
+        /// </summary>
+        /// <example>
+        /// <code>
+        /// var dog = await realm.WriteAsync(() =&gt;
+        /// {
+        ///     return realm.Add(new Dog
+        ///     {
+        ///         Breed = "Dalmatian",
+        ///     });
+        /// });
+        /// </code>
+        /// </example>
+        /// <param name="function">
+        /// Delegate with one return value to execute inside a <see cref="Transaction"/>, creating, updating, or removing objects.
+        /// </param>
+        /// <typeparam name="T">The type returned by the input delegate.</typeparam>
+        /// <returns>An awaitable <see cref="Task"/> with return type <typeparamref name="T"/>.</returns>
+        public async Task<T> WriteAsync<T>(Func<T> function)
+        {
+            ThrowIfDisposed();
+            Argument.NotNull(function, nameof(function));
+
+            // If running on background thread, execute synchronously.
+            if (!AsyncHelper.HasValidContext)
+            {
+                return Write(function);
+            }
+
+            using var transaction = await BeginWriteAsync();
+            var result = function();
+            await transaction.CommitAsync();
+            return result;
         }
 
         /// <summary>
@@ -928,6 +1043,7 @@ namespace Realms
         /// </param>
         /// <typeparam name="T">The type of data in the <see cref="IQueryable{T}"/>.</typeparam>
         /// <returns>An awaitable <see cref="Task"/> with return type <see cref="IQueryable{T}"/>.</returns>
+        [Obsolete("Use Realm.WriteAsync(Func<T> function) instead.")]
         public async Task<IQueryable<T>> WriteAsync<T>(Func<Realm, IQueryable<T>> function)
             where T : RealmObjectBase
         {
@@ -1001,6 +1117,7 @@ namespace Realms
         /// </param>
         /// <typeparam name="T">The type of data in the <see cref="IList{T}"/>.</typeparam>
         /// <returns>An awaitable <see cref="Task"/> with return type <see cref="IList{T}"/>.</returns>
+        [Obsolete("Use Realm.WriteAsync(Func<T> function) instead.")]
         public async Task<IList<T>> WriteAsync<T>(Func<Realm, IList<T>> function)
         {
             ThrowIfDisposed();